refactor(dashboard): extract Court orgs sub-router (F-B-202 PR-5/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-202.md

Sprint 2 modularization PR-5 of 10. base=main.

Extracts 11 routes (the biggest single-PR extract so far) covering
the entire organization management surface:

  app/dashboard/orgs_routes.py (new, 441 LOC, 11 routes):
    GET  /dashboard/orgs                          list (orgs table + invites)
    POST /dashboard/invites/generate              mint admin invite
    POST /dashboard/orgs/{id}/attach-invite       mint attach-CA invite per-org
    POST /dashboard/invites/{id}/revoke           revoke invite
    POST /dashboard/orgs/{id}/approve             approve (sealed-gate)
    POST /dashboard/orgs/{id}/reject              reject (sealed-gate)
    POST /dashboard/orgs/{id}/suspend             suspend (sealed-gate)
    POST /dashboard/orgs/{id}/delete              delete + cascade (sealed-gate)
    POST /dashboard/orgs/{id}/unlock-ca           clear ca_locked (sealed-gate)
    GET  /dashboard/orgs/{id}/upload-ca           upload-CA form
    POST /dashboard/orgs/{id}/upload-ca           upload-CA submit (sealed-gate)

  app/dashboard/router.py (2372 → 2071 LOC, -301):
    - 11 routes removed; breadcrumb comments at the old positions.
    - Sub-router wired via router.include_router(_orgs_routes.router).

All state-changing routes preserve CSRF + ``_require_sealed_reauth``
(audit F-B-2 sealed-org guard) and ``_sealed_mutation_details``
audit logging, imported from ``app/dashboard/_helpers.py``.

Behavioural delta: zero. URLs identical, audit log entries identical,
sealed-gate semantics identical.

Tests:
  - pytest tests/test_dashboard*.py -q — 202/203 (1 pre-existing deselect)
  - pytest tests/ <full suite> -q — 4215 passed, 9 skipped, 32 xfailed
    (2 xdist shared-state flakes pass isolated, pre-existing per
    memory feedback_legacy_sandbox_smoke_obsolete)
  - Stack smoke not run — modern stack exercises Mastio/Connector/
    Frontdesk, not the Court /dashboard surface. Refactor is mechanical
    (extract + include_router, same pattern as PR-1..PR-4).

Router LOC progress:
  Pre-Sprint 2 : 67 routes, 3125 LOC
  Post-PR-1 #839: 67 routes, 2930 LOC
  Post-PR-2 #841: 60 routes, 2614 LOC
  Post-PR-3 #842: 57 routes, 2528 LOC
  Post-PR-4 #847: 50 routes, 2372 LOC
  Post-PR-5 (this): 39 routes, 2071 LOC
  Target post-PR-10: ~3 (aggregator), ~100 LOC

Past the halfway mark on route count (39/67 remaining) and on LOC
(2071/3125 remaining, 34% trimmed).

Next PR-6 (F-B-202 modularization): extract org_onboard_routes.py
+ org_seal_routes.py (sections 18+20, ~325 LOC, 8 routes).
